Fix ClassCastExceptin in Cloud settings panel

+++ epd-shore/src/main/java/dk/dma/epd/shore/gui/settingtabs/ShoreCloudSettingsPanel.java
@@ -122,8 +122,8 @@ public class ShoreCloudSettingsPanel extends CloudSettingsPanel {
         enavSettings.setShoreId(txtShoreId.getText());
         enavSettings.setShorePos(
                 new LatLonPoint.Double(
-                        (Double)spnShoreLat.getValue(), 
-                        (Double)spnShoreLon.getValue()));
+                        ((Float)spnShoreLat.getValue()).doubleValue(), 
+                        ((Float)spnShoreLon.getValue()).doubleValue()));
     }
     
     /**
